Move array creation to edit function
Before this change, the property name array had to be reconstructed in
each manager function. This was not ideal as it led to unnecessary
repetition and complexity.

With this commit, the property name array is now only created in the
edit function. Although the array is created each time the edit
function is called, it significantly reduces repetition across
different manager functions.

The decision to move the array creation to the edit function was made
to avoid repetition and simplify the code in each manager function.