Use an option parameter while deleting a task
Refactored the setUpDeleteClickEvent function to accommodate a new
parameter called displayOption. Previously, this function only accepted
an icon, an id, and a dateSelected parameter, making it tightly coupled
with displaying tasks for a specific date only. With the addition of the
displayOption parameter, the function gains flexibility to handle
different display contexts, such as Date Groups or Project Names.

Now, when calling this function, we can specify whether we want to
display tasks for a Date Group or a Project. Inside the event listener,
the function dynamically determines the display context by checking
whether the provided displayOption matches any Date Group names.
If it does, it calls the displayDayTasks function with the displayOption
parameter; otherwise, it calls the displayProjectTasks function.

This change enhances the modularity and reusability of the code, as it
decouples the display logic from the event setup function and allows
for more versatile usage in different contexts.